rest api added for orders

+++ src/main/java/com/gmail/liliyayalovchenko/dao/EmployeeDAO.java
@@ -8,7 +8,7 @@ public interface EmployeeDAO {
 
     void save(Employee employee);
 
-    Employee getById(Long id);
+    Employee getById(int id);
 
     List<Employee> findAll();
